requirements.txt
flask
pyngrok
google-auth
google-auth-oauthlib
google-api-python-client
gradio
pandas
python-dotenv
google-generativeai
xlsxwriter
plotly